# Conflicts:
#	src/main/java/edu/ntnu/idatt2106/boco/controllers/RatingController.java

+++ src/main/java/edu/ntnu/idatt2106/boco/models/Rating.java
@@ -22,7 +22,7 @@ public class Rating {
     private Long ratingId;
 
     @Column(name = "rate")
-    private int rate;
+    private double rate;
 
     @Column(name = "feedback")
     private String feedback;
@@ -36,7 +36,7 @@ public class Rating {
     private User user;
 
 
-    public Rating(int rate, String feedback, Rental rental, User user){
+    public Rating(double rate, String feedback, Rental rental, User user){
         this.rate = rate;
         this.feedback = feedback;
         this.rental = rental;
